CameraNext: dynamically generate available photo resolutions
Instead of depending on a large matching table, generated the list
of picture sizes to show to the user directly from the supported
list returned by the camera. The list is filtered to remove
uselessly small resolutions on modern devices, to group resolutions
by aspect ratio, and to filter out fairly similar sizes.

Change-Id: I47a67a89786543baec133cf7e71df9819793ebac